Handle "In Progress - Critical" bug
If a game was near the end, it would show in the API as "In Progress - Critical" in the `detailedGameState` property. Because of the matching I was doing, that kept the time remaining from showing. This instead relies on the `abstractGameState`, which supposedly can only be `Preview`, `Live` or `Final`.